requirements.txt
Django==1.6.8
Django-Select2==4.2.2
Pillow==2.7.0
South==1.0.2
argparse==1.2.1
beautifulsoup4==4.3.2
django-ckeditor==4.4.4
django-classy-tags==0.4
django-crispy-forms==1.4.0
django-filer==0.9.5
django-mptt==0.6.0
django-polymorphic==0.6.1
django-reversion==1.8.5
django-sekizai==0.7
django-taggit==0.12.2
djangorestframework==3.1.0
easy-thumbnails==2.2
html5lib==1.0b1
lxml==3.4.1
pi-blogging==0.1.0b1
selenium==2.44.0
six==1.3.0
wsgiref==0.1.2
